perf: split Administration and Settings pages
