+++ app/ui-admin/client/adminFlex.html
@@ -28,31 +28,6 @@
 					{{/if}}
 				{{/each}}
 			</ul>
-
-			{{#if hasSettingPermission}}
-				<h3 class="rooms-list__type">{{_ "Settings"}}</h3>
-				<div class="rc-input sidebar-flex__search">
-					<label class="rc-input__label">
-						<div class="rc-input__wrapper">
-							<div class="rc-input__icon">
-								{{> icon block="rc-input__icon-sv" icon="magnifier"}}
-							</div>
-							<input type="text" class="rc-input__element rc-input__element--small" name="settings-search" placeholder="{{_ 'Search'}}">
-						</div>
-					</label>
-				</div>
-				<ul class="rooms-list__list">
-					{{#each groups}}
-						{{> sidebarItem menuItem label false "admin" _id }}
-					{{else}}
-						<div class="rc-input sidebar-flex__search">
-							<div class="rc-input__wrapper">
-								<p>{{_ "Nothing_found"}}.</p>
-							</div>
-						</div>
-					{{/each}}
-				</ul>
-			{{/if}}
 		</div>
 	</aside>
 </template>
