+++ app/ui-settings/client/index.js
@@ -1,0 +1,2 @@
+import './settingsFlex.html';
+import './settingsFlex';
