+++ client/importPackages.js
@@ -74,6 +74,7 @@ import '../app/tokenpass/client';
 import '../app/ui';
 import '../app/ui-account';
 import '../app/ui-admin/client';
+import '../app/ui-settings/client';
 import '../app/ui-clean-history';
 import '../app/ui-flextab';
 import '../app/ui-login';
